Decompose src/index.ts runtime into explicit controller modules.
Split continuation scheduling, goal persistence, status refresh, and event wiring into named modules so the extension entrypoint only composes controllers.
