+++ src/runtime-config.ts
@@ -1,0 +1,7 @@
+export const CONTINUATION_RETRY_MS = 50;
+export const RUNTIME_PERSIST_INTERVAL_MS = 60_000;
+
+export const __testHooks = {
+  continuationRetryMs: CONTINUATION_RETRY_MS,
+  runtimePersistIntervalMs: RUNTIME_PERSIST_INTERVAL_MS,
+};
